+++ NEWS.md
@@ -1,3 +1,9 @@
-# hellmer 0.1.0
+# hellmer 0.1.1
 
+## New features
 * Initial CRAN submission.
+* Removed `structured_data()` method as `texts()` now handles structured data responses.
+* Updated documentation for better organization and clarity.
+
+## Experimental features
+* Structured data extractions support LLM-as-a-judge to refine extracted data via the `judgements` parameter.